+++ .storybook/helper_classes/index.scss
@@ -94,7 +94,7 @@
 }
 
 .demo-only-spacing-push-background {
-  background: $magenta-light;
+  background: $magenta-base;
 }
 
 .demo-only-spacing-soft-background {
